Fix half/bf16 to float casts for torch JIT compatibility
Use __bfloat162float() and __half2float() intrinsics instead of C-style
(float) casts. The torch JIT compiler defines __CUDA_NO_HALF_CONVERSIONS__
and __CUDA_NO_BFLOAT16_CONVERSIONS__ which disable implicit conversion
operators, breaking (float) casts on these types.

+++ csrc/mla_rope_quant_kernel.cuh
@@ -58,7 +58,7 @@ struct vec_cast<float, nv_bfloat16> {
   template <size_t vec_size>
   MLA_ROPE_INLINE static void cast(float* dst, const nv_bfloat16* src) {
     if constexpr (vec_size == 1) {
-      dst[0] = (float)src[0];
+      dst[0] = __bfloat162float(src[0]);
     } else {
 #pragma unroll
       for (size_t i = 0; i < vec_size / 2; ++i) {
@@ -88,7 +88,7 @@ struct vec_cast<float, half> {
   template <size_t vec_size>
   MLA_ROPE_INLINE static void cast(float* dst, const half* src) {
     if constexpr (vec_size == 1) {
-      dst[0] = (float)src[0];
+      dst[0] = __half2float(src[0]);
     } else {
 #pragma unroll
       for (size_t i = 0; i < vec_size / 2; ++i) {
